refactor(coding-agent): extract turn preparation and action records
Refs ENG-5938; fixes #2185
