+++ packages/coding-agent/src/core/agent-session.ts
@@ -46,12 +46,40 @@ import {
 } from "../session/bash.js";
 import { SessionCommitFence, type SessionCommitLease } from "../session/commit-fence.js";
 import { SessionInputScheduler } from "../session/input-scheduler.js";
+import {
+	buildPromptContent,
+	cloneCustomMessage,
+	cloneQueuedAgentMessage,
+	createDeliveryRecord,
+	createPreparedTurnAction,
+	createSessionCommandAction,
+	DeferredSessionInputError,
+	normalizeMessageContent,
+	type PreparedCommandPayload,
+	type PreparedPromptPreparation,
+	type PreparedTurnPayload,
+	primaryDeliveryRecord,
+	type QueuedAgentMessage,
+	type QueuedSessionAction,
+	queuedAgentMessagePreview,
+	type RestoredPromptInput,
+	SESSION_ACTION_RECOVERY_FORMAT_VERSION,
+	type SessionActionRecoverySnapshot,
+	SessionInputAdmissionPausedError,
+	type SessionInputSchedule,
+	visibleSessionActionProjection,
+} from "../session/prepared-actions.js";
+import {
+	createTurnExecutionPolicy,
+	type TurnExecutionPolicy,
+	TurnPreparer,
+	turnExecutionPoliciesEqual,
+} from "../session/turn-preparation.js";
 import { stripFrontmatter } from "../utils/frontmatter.js";
 import { sleep } from "../utils/sleep.js";
 import { waitForPromiseOrAbort } from "../utils/wait-for-abort.js";
 import {
 	AGENT_MESSAGE_CUSTOM_TYPE,
-	AGENT_MESSAGE_RECEIVED_PREVIEW_LABEL,
 	AGENT_MESSAGE_SKILL_NAME,
 	type AgentFamilyCatalogEntry,
 	type AgentSessionMessage,
@@ -176,7 +204,6 @@ import type { McpManager } from "./mcp/mcp-manager.js";
 import {
 	ASYNC_BASH_COMPLETION_CUSTOM_TYPE,
 	ASYNC_BASH_COMPLETION_PREVIEW_LABEL,
-	type AsyncBashCompletionDetails,
 	type CompactionOutcome,
 	type CompactionOutcomeReason,
 	type CustomMessage,
@@ -281,10 +308,7 @@ import {
 	type RuntimeActivity,
 	type SessionAction,
 	type SessionActionSnapshot,
-	type SessionCommandPayload,
-	type SessionTurnPayload,
 	transitionSessionAction,
-	type WakePolicy,
 } from "./session-action-store.js";
 import type {
 	BranchSummaryEntry,
@@ -409,6 +433,16 @@ export type AgentSessionEvent =
 	| SessionBashEvent
 	| { type: "refine_complete"; result: RefinementResult }
 	| { type: "refine_failed"; error: string };
+
+export {
+	SESSION_ACTION_RECOVERY_FORMAT_VERSION,
+	type SessionActionRecoveryAction,
+	type SessionActionRecoveryPayload,
+	type SessionActionRecoveryRecord,
+	type SessionActionRecoverySnapshot,
+} from "../session/prepared-actions.js";
+
+export type { TurnExecutionPolicy } from "../session/turn-preparation.js";
 
 export type AgentSessionEventListener = (event: AgentSessionEvent) => void;
 
@@ -573,80 +607,6 @@ type NormalizedSubmission =
 	| { kind: "extensionCommand"; completion: Promise<void> }
 	| { kind: "handled" };
 
-type PreTurnCompactionTiming = "beforeModelSelection" | "afterModelSelection" | "skip";
-type RefineBarrierPolicy = "always" | "ifInFlight" | "skip";
-
-interface CommitPreparationPolicy {
-	initialRefineBarrier: RefineBarrierPolicy;
-	flushPendingBashBeforeValidation: boolean;
-	validateModelAndAuth: boolean;
-	awaitPendingModelSelection: boolean;
-	preTurnCompaction: PreTurnCompactionTiming;
-	finalRefineBarrier: RefineBarrierPolicy;
-}
-
-interface CommitPreparationSteps<TPrepared, TCommitted> {
-	afterValidation?: () => void;
-	prepare: () => Promise<TPrepared>;
-	shouldCommit?: (prepared: TPrepared) => boolean;
-	beforeFinalRefineBarrier?: (prepared: TPrepared) => void;
-	commit: (prepared: TPrepared, passedFinalRefineBarrier: boolean) => TCommitted;
-}
-
-type QueuedAgentMessage = UserMessage | CustomMessage;
-type SessionInputSchedule = "steer" | "followUp";
-
-export interface TurnExecutionPolicy {
-	preparation: CommitPreparationPolicy;
-	runBeforeAgentStart: boolean;
-	nextTurnContextTiming: "preparation" | "commit" | "skip";
-	preserveEmptyExtensionPrompt: boolean;
-	completionIncludesRetryChain: boolean;
-}
-
-function turnExecutionPoliciesEqual(left: TurnExecutionPolicy, right: TurnExecutionPolicy): boolean {
-	return (
-		left.preparation.initialRefineBarrier === right.preparation.initialRefineBarrier &&
-		left.preparation.flushPendingBashBeforeValidation === right.preparation.flushPendingBashBeforeValidation &&
-		left.preparation.validateModelAndAuth === right.preparation.validateModelAndAuth &&
-		left.preparation.awaitPendingModelSelection === right.preparation.awaitPendingModelSelection &&
-		left.preparation.preTurnCompaction === right.preparation.preTurnCompaction &&
-		left.preparation.finalRefineBarrier === right.preparation.finalRefineBarrier &&
-		left.runBeforeAgentStart === right.runBeforeAgentStart &&
-		left.nextTurnContextTiming === right.nextTurnContextTiming &&
-		left.preserveEmptyExtensionPrompt === right.preserveEmptyExtensionPrompt &&
-		left.completionIncludesRetryChain === right.completionIncludesRetryChain
-	);
-}
-
-interface PreparedTurnPayload extends SessionTurnPayload {
-	images?: ImageContent[];
-	content?: (TextContent | ImageContent)[];
-	customMessage?: CustomMessage;
-	prepared?: PreparedPromptPreparation;
-	executionPolicy: TurnExecutionPolicy;
-	queueVisible: boolean;
-	acceptedAgentMessage: boolean;
-	acceptedBeforeCompletion: boolean;
-	captureRunMessages?: Set<AgentMessage>;
-	cancelledDispatchEnded?: boolean;
-}
-
-interface PreparedCommandPayload extends SessionCommandPayload {
-	images?: ImageContent[];
-}
-
-type QueuedSessionAction = SessionAction<PreparedTurnPayload | PreparedCommandPayload>;
-
-interface PreparedPromptPreparation {
-	result: Awaited<ReturnType<ExtensionRunner["emitBeforeAgentStart"]>>;
-	basePromptSnapshot: string;
-}
-
-class DeferredSessionInputError extends Error {}
-
-class SessionInputAdmissionPausedError extends Error {}
-
 function oncePreflight(
 	preflightResult: ((success: boolean, queued?: boolean) => void) | undefined,
 ): (success: boolean, queued?: boolean) => void {
@@ -657,121 +617,6 @@ function oncePreflight(
 			preflightResult?.(success, queued);
 		}
 	};
-}
-
-interface RestoredPromptInput {
-	text: string;
-	content?: (TextContent | ImageContent)[];
-	images?: ImageContent[];
-	queueKey?: string;
-	agentMessageId?: string;
-	customMessage?: CustomMessage;
-	prefixMessages?: CustomMessage[];
-}
-
-export const SESSION_ACTION_RECOVERY_FORMAT_VERSION = 1;
-
-export interface SessionActionRecoveryRecord {
-	id: string;
-	role: DeliveryRecord["role"];
-	message: QueuedAgentMessage;
-	ownerActionId: string;
-}
-
-export type SessionActionRecoveryPayload =
-	| {
-			kind: "turn";
-			text: string;
-			preview?: string;
-			records: SessionActionRecoveryRecord[];
-			images?: ImageContent[];
-			content?: (TextContent | ImageContent)[];
-			customMessage?: CustomMessage;
-			executionPolicy: TurnExecutionPolicy;
-			queueVisible: boolean;
-			acceptedAgentMessage: boolean;
-			acceptedBeforeCompletion: boolean;
-	  }
-	| {
-			kind: "session_command";
-			text: string;
-			command: SessionSlashCommand;
-			images?: ImageContent[];
-	  };
-
-export interface SessionActionRecoveryAction {
-	id: string;
-	source: InputSource | "internal";
-	delivery: DeliveryPolicy;
-	wake: WakePolicy;
-	payload: SessionActionRecoveryPayload;
-	queueKey?: string;
-	agentMessageId?: string;
-	suppressAutonomousContinuation?: boolean;
-}
-
-export interface SessionActionRecoverySnapshot {
-	formatVersion: typeof SESSION_ACTION_RECOVERY_FORMAT_VERSION;
-	actions: SessionActionRecoveryAction[];
-}
-
-function cloneCustomMessage(message: CustomMessage): CustomMessage {
-	return {
-		...message,
-		content: Array.isArray(message.content) ? message.content.map((block) => ({ ...block })) : message.content,
-	};
-}
-
-function cloneQueuedAgentMessage(message: QueuedAgentMessage): QueuedAgentMessage {
-	if (message.role === "custom") return cloneCustomMessage(message);
-	return {
-		...message,
-		content: Array.isArray(message.content) ? message.content.map((block) => ({ ...block })) : message.content,
-	};
-}
-
-function primaryDeliveryRecord(action: QueuedSessionAction): DeliveryRecord {
-	if (action.payload.kind !== "turn") throw new Error(`Session action ${action.id} is not a turn`);
-	const record = action.payload.records.find((candidate) => candidate.role === "primary");
-	if (!record) throw new Error(`Turn action ${action.id} has no primary delivery record`);
-	return record;
-}
-
-function normalizeMessageContent(content: string | (TextContent | ImageContent)[]): {
-	text: string;
-	images?: ImageContent[];
-} {
-	if (typeof content === "string") return { text: content };
-	const text = content
-		.filter((part): part is TextContent => part.type === "text")
-		.map((part) => part.text)
-		.join("\n");
-	const images = content.filter((part): part is ImageContent => part.type === "image");
-	return { text, ...(images.length > 0 ? { images } : {}) };
-}
-
-function queuedAgentMessagePreview(action: QueuedSessionAction): string {
-	const payload = action.payload;
-	if (payload.kind === "session_command") return payload.text;
-	if (payload.customMessage && isAgentSessionMessage(payload.customMessage)) {
-		return `${AGENT_MESSAGE_RECEIVED_PREVIEW_LABEL}: ${payload.customMessage.details.message}`;
-	}
-	if (payload.customMessage?.customType === ASYNC_BASH_COMPLETION_CUSTOM_TYPE) {
-		const details = payload.customMessage.details as AsyncBashCompletionDetails | undefined;
-		return details
-			? `${ASYNC_BASH_COMPLETION_PREVIEW_LABEL}: pid ${details.pid}, exit ${details.exitCode}`
-			: ASYNC_BASH_COMPLETION_PREVIEW_LABEL;
-	}
-	return payload.preview ?? payload.text;
-}
-
-function visibleSessionActionProjection(actions: readonly QueuedSessionAction[]): readonly QueuedSessionAction[] {
-	return actions.filter(
-		(action) =>
-			action.payload.kind === "session_command" ||
-			action.payload.queueVisible ||
-			action.payload.acceptedAgentMessage,
-	);
 }
 
 const IPYTHON_SENT_AGENT_MESSAGE_CUSTOM_ENTRY = "ipython_sent_agent_message";
@@ -1186,6 +1031,14 @@ export class AgentSession {
 	private _sessionInputArrivalEpoch = 0;
 	private readonly _durableRlmTerminalNoticeActionIds = new Set<string>();
 	private readonly _commitFence = new SessionCommitFence();
+	private readonly _turnPreparer = new TurnPreparer({
+		hasRefinement: () => this._refineInFlight !== undefined,
+		waitForRefinement: () => this._waitForRefineIdle(),
+		flushPendingBash: () => this._flushPendingBashMessages(),
+		validate: () => this._validateCanStartAgentRun(),
+		compact: () => this._runPreTurnCompaction(),
+		pendingModelSelection: () => this._pendingModelSelectEmit(),
+	});
 	// Checkpoint, handoff, and activity waiters share lifecycle-edge notifications to avoid polling.
 	private readonly _sessionInputCheckpointWaiters = new Set<() => void>();
 	private _pendingNextTurnMessages: CustomMessage[] = [];
@@ -4366,42 +4219,6 @@ export class AgentSession {
 		if (lastAssistant) await this._checkCompaction(lastAssistant, false, false);
 	}
 
-	private async _prepareForCommit<TPrepared, TCommitted>(
-		policy: CommitPreparationPolicy,
-		steps: CommitPreparationSteps<TPrepared, TCommitted>,
-	): Promise<TCommitted | undefined> {
-		if (
-			policy.initialRefineBarrier === "always" ||
-			(policy.initialRefineBarrier === "ifInFlight" && this._refineInFlight)
-		) {
-			await this._waitForRefineIdle();
-		}
-		if (policy.flushPendingBashBeforeValidation) this._flushPendingBashMessages();
-		if (policy.validateModelAndAuth) await this._validateCanStartAgentRun();
-		steps.afterValidation?.();
-		if (!policy.flushPendingBashBeforeValidation) this._flushPendingBashMessages();
-
-		if (policy.preTurnCompaction === "beforeModelSelection") await this._runPreTurnCompaction();
-		if (policy.awaitPendingModelSelection) {
-			const pendingModelSelectEmit = this._pendingModelSelectEmit();
-			if (pendingModelSelectEmit) await pendingModelSelectEmit;
-		}
-		if (policy.preTurnCompaction === "afterModelSelection") await this._runPreTurnCompaction();
-
-		const prepared = await steps.prepare();
-		if (steps.shouldCommit && !steps.shouldCommit(prepared)) return undefined;
-		steps.beforeFinalRefineBarrier?.(prepared);
-		let passedFinalRefineBarrier = false;
-		if (
-			policy.finalRefineBarrier === "always" ||
-			(policy.finalRefineBarrier === "ifInFlight" && this._refineInFlight)
-		) {
-			await this._waitForRefineIdle();
-			passedFinalRefineBarrier = true;
-		}
-		return steps.commit(prepared, passedFinalRefineBarrier);
-	}
-
 	private _applyPreparedSystemPrompt(
 		preparation: PreparedPromptPreparation | undefined,
 		preserveEmptyExtensionPrompt: boolean,
@@ -4579,7 +4396,7 @@ export class AgentSession {
 			suppressAutonomousContinuation: true,
 			resumeIfIdle: false,
 			source: "internal",
-			executionPolicy: this._turnExecutionPolicy("injected"),
+			executionPolicy: createTurnExecutionPolicy("injected"),
 			queueVisible: false,
 		});
 		this._durableRlmTerminalNoticeActionIds.add(action.id);
@@ -4751,7 +4568,7 @@ export class AgentSession {
 				source: options?.source ?? "internal",
 				executionPolicy:
 					options?.executionPolicy ??
-					(visibleQueued ? this._turnExecutionPolicy("queued") : this._turnExecutionPolicy("injected")),
+					(visibleQueued ? createTurnExecutionPolicy("queued") : createTurnExecutionPolicy("injected")),
 				queueVisible: visibleQueued,
 			});
 			const result = this._admitSessionInput(action, {
@@ -4844,7 +4661,7 @@ export class AgentSession {
 				const wasBusy = wasRuntimeBusy || pendingOwnedWork;
 				if (normalized.kind === "sessionCommand") {
 					const schedule = options?.streamingBehavior ?? (this.isStreaming ? "steer" : "followUp");
-					const action = this._createSessionCommandAction(
+					const action = createSessionCommandAction(
 						normalized.text,
 						normalized.command,
 						normalized.images,
@@ -4882,7 +4699,7 @@ export class AgentSession {
 				const prefixMessages = visibleQueued ? this._takePendingNextTurnMessages() : undefined;
 				const content = options?.content
 					? options.content.map((block) => ({ ...block }))
-					: this._buildPromptContent(normalized.text, normalized.images);
+					: buildPromptContent(normalized.text, normalized.images);
 				const suppliedMessage = options?.customMessage;
 				const primaryMessage = suppliedMessage
 					? visibleQueued
@@ -4907,8 +4724,8 @@ export class AgentSession {
 						(options?.queueIfBusy === true && canSelectSessionAction(this._runtimeActivity())),
 					source: isInternalPrompt ? "internal" : (options?.source ?? "interactive"),
 					executionPolicy: visibleQueued
-						? this._turnExecutionPolicy("queued")
-						: this._turnExecutionPolicy("directPrompt", {
+						? createTurnExecutionPolicy("queued")
+						: createTurnExecutionPolicy("directPrompt", {
 								returnAfterAccepted: options?.returnAfterAccepted,
 								skipPrePromptWork: options?.skipPrePromptWork,
 							}),
@@ -5210,7 +5027,7 @@ export class AgentSession {
 			return undefined;
 		}
 		return this._admitSessionInput(
-			this._createSessionCommandAction(text, customMessage.details.command, images, schedule, {
+			createSessionCommandAction(text, customMessage.details.command, images, schedule, {
 				agentMessageId,
 				source: "internal",
 			}),
@@ -5287,193 +5104,14 @@ export class AgentSession {
 		});
 	}
 
-	private _buildPromptContent(text: string, images?: ImageContent[]): (TextContent | ImageContent)[] {
-		const content: (TextContent | ImageContent)[] = [];
-		content.push({ type: "text", text });
-		if (images) content.push(...images);
-		return content;
-	}
-
 	private _takePendingNextTurnMessages(): CustomMessage[] {
 		const messages = this._pendingNextTurnMessages;
 		this._pendingNextTurnMessages = [];
 		return messages;
 	}
 
-	private _deliveryPolicy(schedule: SessionInputSchedule): DeliveryPolicy {
-		return schedule === "steer" ? "next_turn_boundary" : "when_run_idle";
-	}
-
-	private _createDeliveryRecord(
-		actionId: string,
-		role: DeliveryRecord["role"],
-		message: QueuedAgentMessage,
-	): DeliveryRecord {
-		return {
-			id: randomUUID(),
-			role,
-			message,
-			started: false,
-			durable: false,
-			ownerActionId: actionId,
-		};
-	}
-
-	private _turnExecutionPolicy(
-		kind: "queued" | "directPrompt" | "injected" | "customTrigger",
-		options: {
-			returnAfterAccepted?: boolean;
-			skipPrePromptWork?: boolean;
-		} = {},
-	): TurnExecutionPolicy {
-		if (kind === "queued") {
-			return {
-				preparation: {
-					initialRefineBarrier: "skip",
-					flushPendingBashBeforeValidation: false,
-					validateModelAndAuth: true,
-					awaitPendingModelSelection: true,
-					preTurnCompaction: "beforeModelSelection",
-					finalRefineBarrier: "always",
-				},
-				runBeforeAgentStart: true,
-				nextTurnContextTiming: "commit",
-				preserveEmptyExtensionPrompt: true,
-				completionIncludesRetryChain: true,
-			};
-		}
-		if (kind === "directPrompt") {
-			return {
-				preparation: {
-					initialRefineBarrier: options.returnAfterAccepted ? "skip" : "always",
-					flushPendingBashBeforeValidation: true,
-					validateModelAndAuth: true,
-					awaitPendingModelSelection: true,
-					preTurnCompaction: options.skipPrePromptWork ? "skip" : "afterModelSelection",
-					finalRefineBarrier: "ifInFlight",
-				},
-				runBeforeAgentStart: !options.skipPrePromptWork,
-				nextTurnContextTiming: "preparation",
-				preserveEmptyExtensionPrompt: false,
-				completionIncludesRetryChain: true,
-			};
-		}
-		if (kind === "injected") {
-			return {
-				preparation: {
-					initialRefineBarrier: "always",
-					flushPendingBashBeforeValidation: true,
-					validateModelAndAuth: true,
-					awaitPendingModelSelection: true,
-					preTurnCompaction: "beforeModelSelection",
-					finalRefineBarrier: "ifInFlight",
-				},
-				runBeforeAgentStart: true,
-				nextTurnContextTiming: "preparation",
-				preserveEmptyExtensionPrompt: true,
-				completionIncludesRetryChain: true,
-			};
-		}
-		return {
-			preparation: {
-				initialRefineBarrier: "always",
-				flushPendingBashBeforeValidation: false,
-				validateModelAndAuth: false,
-				awaitPendingModelSelection: false,
-				preTurnCompaction: "skip",
-				finalRefineBarrier: "skip",
-			},
-			runBeforeAgentStart: false,
-			nextTurnContextTiming: "skip",
-			preserveEmptyExtensionPrompt: false,
-			completionIncludesRetryChain: false,
-		};
-	}
-
-	private _createPreparedTurnAction(
-		schedule: SessionInputSchedule,
-		text: string,
-		images: ImageContent[] | undefined,
-		options: {
-			agentMessageId?: string;
-			queueKey?: string;
-			content?: (TextContent | ImageContent)[];
-			message?: QueuedAgentMessage;
-			prefixMessages?: CustomMessage[];
-			previewLabel?: string;
-			suppressAutonomousContinuation?: boolean;
-			resumeIfIdle?: boolean;
-			source?: InputSource | "internal";
-			executionPolicy?: TurnExecutionPolicy;
-			queueVisible?: boolean;
-			acceptedAgentMessage?: boolean;
-			acceptedBeforeCompletion?: boolean;
-		},
-	): QueuedSessionAction {
-		const id = randomUUID();
-		const content = options.content ?? this._buildPromptContent(text, images);
-		const message =
-			options.message ??
-			({
-				role: "user",
-				content: content.map((block) => ({ ...block })),
-				timestamp: Date.now(),
-			} satisfies UserMessage);
-		const prefixMessages = options.prefixMessages?.map((prefix) => cloneCustomMessage(prefix)) ?? [];
-		const preview = options.previewLabel ? `${options.previewLabel}: ${text}` : undefined;
-		const payload: PreparedTurnPayload = {
-			kind: "turn",
-			text,
-			records: [
-				...prefixMessages.map((prefix) => this._createDeliveryRecord(id, "prefix", prefix)),
-				this._createDeliveryRecord(id, "primary", message),
-			],
-			preview,
-			images: images?.map((image) => ({ ...image })),
-			content: content.map((block) => ({ ...block })),
-			customMessage: options.message?.role === "custom" ? cloneCustomMessage(options.message) : undefined,
-			executionPolicy: options.executionPolicy ?? this._turnExecutionPolicy("queued"),
-			queueVisible: options.queueVisible ?? true,
-			acceptedAgentMessage: options.acceptedAgentMessage ?? false,
-			acceptedBeforeCompletion: options.acceptedBeforeCompletion ?? false,
-		};
-		return {
-			id,
-			source: options.source ?? "internal",
-			delivery: this._deliveryPolicy(schedule),
-			wake:
-				options.resumeIfIdle === true
-					? "immediate"
-					: schedule === "steer"
-						? "on_lower_boundary"
-						: "external_resume",
-			payload,
-			lifecycle: { state: "queued" },
-			queueKey: options.queueKey,
-			agentMessageId: options.agentMessageId,
-			suppressAutonomousContinuation: options.suppressAutonomousContinuation,
-		};
-	}
-
-	private _createSessionCommandAction(
-		text: string,
-		command: SessionSlashCommand,
-		images: ImageContent[] | undefined,
-		schedule: SessionInputSchedule,
-		options: {
-			agentMessageId?: string;
-			source?: InputSource | "internal";
-		} = {},
-	): QueuedSessionAction {
-		return {
-			id: randomUUID(),
-			source: options.source ?? "internal",
-			delivery: this._deliveryPolicy(schedule),
-			wake: "immediate",
-			payload: { kind: "session_command", text, command, images },
-			lifecycle: { state: "queued" },
-			agentMessageId: options.agentMessageId,
-		};
+	private _createPreparedTurnAction(...args: Parameters<typeof createPreparedTurnAction>): QueuedSessionAction {
+		return createPreparedTurnAction(...args);
 	}
 
 	private _coalescedFollowUpOwner(action: QueuedSessionAction): QueuedSessionAction | undefined {
@@ -5911,7 +5549,7 @@ export class AgentSession {
 			nextTurnMessages = [];
 		};
 		try {
-			const preparedTurn = await this._prepareForCommit(executionPolicy.preparation, {
+			const preparedTurn = await this._turnPreparer.prepare(executionPolicy.preparation, {
 				afterValidation: () => {
 					if (this._isSessionInputHandoffDeferred(epoch)) {
 						throw new DeferredSessionInputError("Session input paused before preflight");
@@ -5983,7 +5621,7 @@ export class AgentSession {
 						}
 					}
 					const contextRecords = nextTurnMessages.map((message) =>
-						this._createDeliveryRecord(turns[0].id, "next_turn", message),
+						createDeliveryRecord(turns[0].id, "next_turn", message),
 					);
 					const firstPrimaryIndex = turns[0].payload.records.indexOf(primaryDeliveryRecord(turns[0]));
 					turns[0].payload.records.splice(firstPrimaryIndex, 0, ...contextRecords);
@@ -6197,7 +5835,7 @@ export class AgentSession {
 				const action = this._createPreparedTurnAction("followUp", normalized.text, normalized.images, {
 					message: appMessage,
 					resumeIfIdle: true,
-					executionPolicy: this._turnExecutionPolicy("customTrigger"),
+					executionPolicy: createTurnExecutionPolicy("customTrigger"),
 					queueVisible: false,
 				});
 				const result = this._admitSessionInput(action, { immediatelyEligible });
